Update 달팽이는 올라가고 싶다.java

+++ 백준/Bronze/2869. 달팽이는 올라가고 싶다/달팽이는 올라가고 싶다.java
@@ -9,7 +9,7 @@ public class Main {
 		int b = sc.nextInt();
 		int v = sc.nextInt();
 		
-		long res = (v - a) / (a -b);
+		int res = (v - a) / (a -b);
 
 		if ((v - a) % (a - b) == 0) {
 			System.out.println(res + 1);
